HZN-1315: Refactor the Kafka Sink impl. into different modules, allowing the client (producer) and server (consumer) to be loaded separately.
