--- core/ipc/sink/kafka-impl/src/main/java/org/opennms/core/ipc/sink/kafka/offset/KafkaOffsetConstants.java
@@ -1,51 +1,0 @@
-/*******************************************************************************
- * This file is part of OpenNMS(R).
- *
- * Copyright (C) 2016 The OpenNMS Group, Inc.
- * OpenNMS(R) is Copyright (C) 1999-2016 The OpenNMS Group, Inc.
- *
- * OpenNMS(R) is a registered trademark of The OpenNMS Group, Inc.
- *
- * OpenNMS(R) is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published
- * by the Free Software Foundation, either version 3 of the License,
- * or (at your option) any later version.
- *
- * OpenNMS(R) is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with OpenNMS(R).  If not, see:
- *      http://www.gnu.org/licenses/
- *
- * For more information contact:
- *     OpenNMS(R) Licensing <license@opennms.org>
- *     http://www.opennms.org/
- *     http://www.opennms.com/
- *******************************************************************************/
-
-package org.opennms.core.ipc.sink.kafka.offset;
-
-public interface KafkaOffsetConstants {
-
-    static final String OFFSET = "offset";
-
-    static final String TOPIC = "topic";
-
-    static final String GROUP = "group";
-
-    static final String PARTITION = "partition";
-
-    static final int TIMEOUT = 100000;
-
-    static final int BUFFERSIZE = 64 * 1024;
-
-    static final int POLL_INTERVAL = 500;
-
-    static final String OFFSETS_TOPIC = "__consumer_offsets";
-    
-    static final String CLIENT_NAME = "OpenNMS-Kafka-Monitor";
-
-}
