--- core/ipc/sink/kafka-impl/src/main/resources/META-INF/opennms/applicationContext-ipc-sink-server-kafka.xml
@@ -1,27 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<beans xmlns="http://www.springframework.org/schema/beans"
-  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xmlns:context="http://www.springframework.org/schema/context"
-  xmlns:util="http://www.springframework.org/schema/util"
-  xmlns:onmsgi="http://xmlns.opennms.org/xsd/spring/onms-osgi"
-  xsi:schemaLocation="
-  http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-4.2.xsd
-  http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context-4.2.xsd
-  http://www.springframework.org/schema/util http://www.springframework.org/schema/util/spring-util-4.2.xsd
-  http://xmlns.opennms.org/xsd/spring/onms-osgi http://xmlns.opennms.org/xsd/spring/onms-osgi.xsd
-">
-
-  <context:annotation-config />
-
-  <bean id="kafkaMessageConsumerManager" class="org.opennms.core.ipc.sink.kafka.KafkaMessageConsumerManager" />
-
-  <onmsgi:service ref="kafkaMessageConsumerManager" interface="org.opennms.core.ipc.sink.api.MessageConsumerManager" />
-
-  <bean id="kafkaLocalMessageDispatcherFactory" class="org.opennms.core.ipc.sink.kafka.KafkaLocalMessageDispatcherFactory" />
-
-  <onmsgi:service ref="kafkaLocalMessageDispatcherFactory" interface="org.opennms.core.ipc.sink.api.MessageDispatcherFactory" />
-
-  <bean id="kafkaOffsetProvider" class="org.opennms.core.ipc.sink.kafka.offset.KafkaOffsetProvider"
-		init-method="start" destroy-method="stop" />
-
-</beans>
