--- core/ipc/sink/kafka-impl/src/main/resources/META-INF/opennms/component-dao.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<beans xmlns="http://www.springframework.org/schema/beans" 
-       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-       xmlns:context="http://www.springframework.org/schema/context"
-       xsi:schemaLocation="
-       http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd
-       http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context-4.2.xsd
-       ">
-
-    <!-- Conditionally load the Kafka Sink -->
-    <bean class="org.opennms.core.ipc.sink.kafka.ConditionalKafkaSinkContext"/>
-
-</beans>
