+++ core/ipc/sink/kafka/itests/src/test/java/org/opennms/core/ipc/sink/kafka/itests/heartbeat/HeartbeatModule.java
@@ -1,0 +1,65 @@
+/*******************************************************************************
+ * This file is part of OpenNMS(R).
+ *
+ * Copyright (C) 2016 The OpenNMS Group, Inc.
+ * OpenNMS(R) is Copyright (C) 1999-2016 The OpenNMS Group, Inc.
+ *
+ * OpenNMS(R) is a registered trademark of The OpenNMS Group, Inc.
+ *
+ * OpenNMS(R) is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published
+ * by the Free Software Foundation, either version 3 of the License,
+ * or (at your option) any later version.
+ *
+ * OpenNMS(R) is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with OpenNMS(R).  If not, see:
+ *      http://www.gnu.org/licenses/
+ *
+ * For more information contact:
+ *     OpenNMS(R) Licensing <license@opennms.org>
+ *     http://www.opennms.org/
+ *     http://www.opennms.com/
+ *******************************************************************************/
+
+package org.opennms.core.ipc.sink.kafka.itests.heartbeat;
+
+import org.opennms.core.ipc.sink.api.AggregationPolicy;
+import org.opennms.core.ipc.sink.api.AsyncPolicy;
+import org.opennms.core.ipc.sink.xml.AbstractXmlSinkModule;
+
+public class HeartbeatModule extends AbstractXmlSinkModule<Heartbeat,Heartbeat> {
+
+    public static final HeartbeatModule INSTANCE = new HeartbeatModule();
+
+    public HeartbeatModule() {
+        super(Heartbeat.class);
+    }
+
+    @Override
+    public int getNumConsumerThreads() {
+        return 1;
+    }
+
+    @Override
+    public String getId() {
+        return "Heartbeat";
+    }
+
+    @Override
+    public AggregationPolicy<Heartbeat, Heartbeat, Heartbeat> getAggregationPolicy() {
+        // No aggregation
+        return null;
+    }
+
+    @Override
+    public AsyncPolicy getAsyncPolicy() {
+        // Only synchronous dispatching
+        return null;
+    }
+
+}
