+++ core/ipc/sink/kafka/itests/src/test/java/org/opennms/core/ipc/sink/kafka/itests/offset/KafkaMessageProducer.java
@@ -1,0 +1,76 @@
+/*******************************************************************************
+ * This file is part of OpenNMS(R).
+ *
+ * Copyright (C) 2016 The OpenNMS Group, Inc.
+ * OpenNMS(R) is Copyright (C) 1999-2016 The OpenNMS Group, Inc.
+ *
+ * OpenNMS(R) is a registered trademark of The OpenNMS Group, Inc.
+ *
+ * OpenNMS(R) is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published
+ * by the Free Software Foundation, either version 3 of the License,
+ * or (at your option) any later version.
+ *
+ * OpenNMS(R) is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with OpenNMS(R).  If not, see:
+ *      http://www.gnu.org/licenses/
+ *
+ * For more information contact:
+ *     OpenNMS(R) Licensing <license@opennms.org>
+ *     http://www.opennms.org/
+ *     http://www.opennms.com/
+ *******************************************************************************/
+
+package org.opennms.core.ipc.sink.kafka.itests.offset;
+
+import java.util.Properties;
+
+import org.apache.kafka.clients.producer.KafkaProducer;
+import org.apache.kafka.clients.producer.Producer;
+import org.apache.kafka.clients.producer.ProducerRecord;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+public class KafkaMessageProducer {
+
+    private static final Logger LOGGER = LoggerFactory.getLogger(KafkaMessageProducer.class);
+
+    private String bootStrapServer;
+
+    public KafkaMessageProducer(String bootStrapServer) {
+        this.bootStrapServer = bootStrapServer;
+    }
+
+    public void produce() {
+        Properties props = new Properties();
+        props.put("bootstrap.servers", getBootStrapServer());
+        props.put("acks", "all");
+        props.put("retries", 0);
+        props.put("batch.size", 16384);
+        props.put("linger.ms", 1);
+        props.put("buffer.memory", 33554432);
+        props.put("key.serializer", "org.apache.kafka.common.serialization.StringSerializer");
+        props.put("value.serializer", "org.apache.kafka.common.serialization.StringSerializer");
+        Producer<String, String> producer = new KafkaProducer<>(props);
+        for (int i = 0; i < 1000; i++) {
+            producer.send(new ProducerRecord<String, String>("USER_TOPIC", Integer.toString(i),
+                    "This is  " + Integer.toString(i) + "  Message"));
+        }
+        LOGGER.info("produced 1000 messages in USER_TOPIC");
+        producer.close();
+    }
+
+    public String getBootStrapServer() {
+        return bootStrapServer;
+    }
+
+    public void setBootStrapServer(String bootStrapServer) {
+        this.bootStrapServer = bootStrapServer;
+    }
+
+}
