+++ core/ipc/sink/kafka/itests/src/test/resources/META-INF/opennms/applicationContext-rpc-echo.xml
