+++ features/minion/repository/pom.xml
@@ -252,8 +252,8 @@
             <version>${project.version}</version>
         </dependency>
         <dependency>
-            <groupId>org.opennms.core.ipc.sink</groupId>
-            <artifactId>org.opennms.core.ipc.sink.kafka-impl</artifactId>
+            <groupId>org.opennms.core.ipc.sink.kafka</groupId>
+            <artifactId>org.opennms.core.ipc.sink.kafka.client</artifactId>
             <version>${project.version}</version>
         </dependency>
         <dependency>
